fuel_routes_api/serializers.py
from rest_framework import serializers
from .models import FuelStation


class FuelStationSerializer(serializers.ModelSerializer):
    class Meta:
        model = FuelStation
        fields = ['name', 'address', 'latitude', 'longitude', 'retail_price']
