requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==3.2.23
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
numpy==1.24.4
pandas==2.0.3
psycopg==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
